TileByteCounts is now an Option[Array[Long]]

+++ raster/src/main/scala/geotrellis/raster/io/geotiff/tags/TileTags.scala
@@ -7,5 +7,5 @@ case class TileTags(
   tileWidth: Option[Long] = None,
   tileLength: Option[Long] = None,
   tileOffsets: Option[Array[Long]] = None,
-  tileByteCounts: Option[Array[Int]] = None
+  tileByteCounts: Option[Array[Long]] = None
 )
